requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_google_genai
langgraph
duckduckgo-search
